CGPoint causes rendering issue on arm64, change data format to GLfloat.

+++ framework/Source/iOS/MultiFaceFilter.h
@@ -3,7 +3,7 @@
 @interface MultiFaceFilter<AVCaptureMetadataOutputObjectsDelegate> : GPUImageFilter
 {
     NSMutableArray *filters;
-    int face_count;
+    NSUInteger face_count;
 }
 
 @property(readwrite, nonatomic) BOOL mirror;
